refactor playground
