+++ packages/playground/src/BotLogger.ts
@@ -1,0 +1,3 @@
+export interface BotLogger {
+  info: (action: string, message: string) => void
+}
