+++ packages/playground/src/bots/FoundationBot.ts
@@ -1,11 +1,11 @@
-import { RegTestGenesisKeys } from '@defichain/jellyfish-network'
+import { RegTestFoundationKeys } from '@defichain/jellyfish-network'
 import { AbstractBot } from '../AbstractBot'
 
 /**
  * Bootstrap Foundation Keys
  */
 export class FoundationBot extends AbstractBot {
-  static Keys = RegTestGenesisKeys
+  static Keys = RegTestFoundationKeys
 
   async bootstrap (): Promise<void> {
     for (const key of FoundationBot.Keys) {
