+++ packages/playground/tsconfig.build.json
@@ -5,6 +5,6 @@
   ],
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": "./src",
+    "rootDir": "./src"
   }
 }
